% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{influenza_germany}
\alias{influenza_germany}
\title{Germany Influenza (FLU) Surveillance data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1037 rows and 4 columns.
}
\source{
License CC-BY 4.0: Robert Koch-Institut (2025): Laborbestätigte Influenzafälle in Deutschland. Dataset. Zenodo.
DOI:10.5281/zenodo.14619502. \url{https://github.com/robert-koch-institut/Influenzafaelle_in_Deutschland}
}
\usage{
influenza_germany
}
\description{
A subset of the weekly German influenza surveillance data from January 2020 to January 2025.
}
\details{
A data frame with 1,037 rows and 4 columns:
\describe{
\item{ReportingWeek}{Reporting Week in "2024-W03" format}
\item{AgeGroup}{Age groups: \verb{00+} for all and \code{00-14}, \code{15-59} and \verb{60+} for age stratified cases.}
\item{Cases}{Weekly case count}
\item{Incidence}{Calculated weekly incidence}
}
}
\keyword{datasets}
