% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_helpers.R
\name{th_disable_legend, th_legend_position}
\alias{th_disable_legend, th_legend_position}
\alias{th_disable_legend}
\alias{th_legend_position}
\alias{th_legend_top}
\alias{th_legend_bottom}
\alias{th_legend_left}
\alias{th_legend_right}
\title{Quickly remove the minor lines of the panel grid}
\usage{
th_disable_legend()

th_legend_position(
  position = c("top", "bottom", "left", "right", "none", "inside"),
  position.inside = NULL
)

th_legend_top()

th_legend_bottom()

th_legend_left()

th_legend_right()
}
\arguments{
\item{position}{Position of the ggplot2 legend.
Options are \verb{("top", "bottom", "left", "right", "none", "inside")}}

\item{position.inside}{Coordinates for the legend inside the plot.
If set overwrites \code{position} to \code{inside}.}
}
\value{
Changes the \code{legend.position} of the \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
Convenience functions to control the legend position for \code{ggplot2}.
Has to be called after setting the theme.
}
