% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_helpers.R
\name{Quick legend helpers}
\alias{Quick legend helpers}
\alias{th_guide_thick_lines}
\alias{th_guide_long_lines}
\alias{th_remove_legend_title}
\title{Quick legend helpers}
\usage{
th_guide_thick_lines(linewidth = 3, ...)

th_guide_long_lines(keywidth = 3, ...)

th_remove_legend_title()
}
\arguments{
\item{linewidth}{Line width of the line shown in the colour legend.}

\item{...}{further parameters passed to \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}.}

\item{keywidth}{Key width of the line shown in the legend.}
}
\value{
Overrides the \code{aes} of the colour guide using \code{\link[ggplot2:guides]{ggplot2::guides()}}.
}
\description{
\itemize{
\item \code{th_guide_thick_lines()}: Increase the line thickness for \code{geom_line()} \strong{colour} legends.
\item \code{th_guide_long_lines()}: Increase the line thickness for \code{geom_line()} \strong{linetype} legends.
\item \code{th_remove_legend_title()}: Removes the legend title. Use after setting the theme.
}
}
