% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_helpers.R
\name{th_remove_panel_grid_minor}
\alias{th_remove_panel_grid_minor}
\alias{th_remove_panel_grid_minor_y}
\alias{th_remove_panel_grid_minor_x}
\alias{th_remove_panel_grid}
\title{Quickly remove the minor lines of the panel grid}
\usage{
th_remove_panel_grid_minor()

th_remove_panel_grid_minor_y()

th_remove_panel_grid_minor_x()

th_remove_panel_grid()
}
\value{
Changes the \code{panel.grid.minor} of the \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
\code{th_remove_panel_grid_minor()}, \code{th_remove_panel_grid_minor_x()},
\code{th_remove_panel_grid_minor_y()} are convenience functions remove the minor lines
of the panel grid.
Has to be called after setting the theme.
}
