% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_helpers.R
\name{theme_mod_remove_minor_grid}
\alias{theme_mod_remove_minor_grid}
\alias{theme_mod_remove_minor_grid_y}
\alias{theme_mod_remove_minor_grid_x}
\alias{theme_mod_remove_panel_grid}
\title{Quickly remove the minor lines of the panel grid}
\usage{
theme_mod_remove_minor_grid()

theme_mod_remove_minor_grid_y()

theme_mod_remove_minor_grid_x()

theme_mod_remove_panel_grid()
}
\value{
Changes the \code{panel.grid.minor} of the \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
\code{theme_mod_remove_minor_grid()}, \code{theme_mod_remove_minor_grid_x()},
\code{theme_mod_remove_minor_grid_y()} are convenience functions remove the minor lines
of the panel grid.
Has to be called after setting the theme.
}
