% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linelist_hospital_outbreak}
\alias{linelist_hospital_outbreak}
\title{Line list of a fictional hospital outbreak}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 9 columns.
}
\usage{
linelist_hospital_outbreak
}
\description{
This hospital outbreak is inspired by typical hospital outbreaks with resistant 4MRGN bacterial pathogens.
These outbreaks start silent, since they are not initially apparent from the symptoms of the patient.
}
\details{
A data frame with 8 rows and 9 columns:
\describe{
\item{Patient}{Patient id}
}
}
\keyword{datasets}
