% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tidyr.R
\name{uncount}
\alias{uncount}
\alias{expand_counts}
\title{Duplicate rows according to a weighting variable}
\usage{
expand_counts(data, weights, ..., .remove = TRUE, .id = NULL)
}
\value{
A data frame with rows duplicated according to weights.
}
\description{
\code{uncount()} and its alias \code{expand_counts()} are complements of \code{\link[dplyr:count]{dplyr::count()}}: they take
a data frame with a column of frequencies and duplicate each row according to
those frequencies.

\code{uncount()} is provided by the tidyr package, and re-exported
by ggsurveillance. See \code{\link[tidyr:uncount]{tidyr::uncount()}} for more details.
}
\examples{
df <- data.frame(x = c("a", "b"), n = c(2, 3))
uncount(df, n)
# Or equivalently:
expand_count(df, n)

}
